Move FHIR utils to separate files
