+++ util/MeasureCalculation.ts
@@ -1,6 +1,6 @@
 import { Calculator, CalculatorTypes } from 'fqm-execution';
 import { TestCaseInfo } from '../state/atoms/patientTestCase';
-import { createPatientBundle } from './fhir';
+import { createPatientBundle } from './fhir/resourceCreation';
 
 /**
  * Takes in a patient's test case info, a measure bundle, and a measurement
